Move hyoga.core.download to hyoga.open.downloader.
