+++ hyoga/open/example.py
@@ -10,7 +10,7 @@ series and plotting output from other models.
 
 import xarray as xr
 
-import hyoga.core.download
+import hyoga.open.downloader
 
 
 def example(filename='pism.alps.out.2d.nc'):
@@ -18,5 +18,5 @@ def example(filename='pism.alps.out.2d.nc'):
     repo = 'https://raw.githubusercontent.com/juseg/hyoga-data/main'
     model = filename.split('.')[0]
     url = '/'.join((repo, model, filename))
-    path = hyoga.core.download.Downloader()(url, filename)
+    path = hyoga.open.downloader.Downloader()(url, filename)
     return xr.open_dataset(path)
